project should only be deleted if it exists
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Analyzer/Framework/trunk@29573 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ org.palladiosimulator.analyzer.workflow/src/org/palladiosimulator/analyzer/workflow/jobs/CreatePluginProjectJob.java
@@ -127,11 +127,14 @@ public class CreatePluginProjectJob implements IJob {
 			if (myProject == null) {
 				return;
 			}
-
-			try {
-				deleteProject(monitor, myProject);
-			} catch (CoreException e) {
-				throw new CleanupFailedException("Delete project failed", e);
+			
+			// project should only be deleted if it exists.
+			if(myProject.exists()) {
+			    try {
+			        deleteProject(monitor, myProject);
+			    } catch (CoreException e) {
+			        throw new CleanupFailedException("Delete project failed", e);
+			    }			    
 			}
 		}
 	}
